split out transform class
