+++ __test__/bitmap.test.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const readFile = require('../index');
+const Tranform = require('../lib/transform');
 const image = '../assets/baldy.bmp';
 let x = readFile(image, 'allBlack');
 
@@ -10,7 +11,13 @@ describe ('testing Bitmap functionality', () => {
   });
   test('testing Color Array', () => {
     // const check = readFile(image, 'allBlack');
-    expect(x.colorArray[0]).toEqual(0);
+    let testColorArray = readFile('../assets/baldy.bmp', 'allBlack');
+    console.log(testColorArray);
+    expect(testColorArray[0]).toEqual(0);
   });
+  test('testing negative transform', () => {
+    const testArray = [1,2,3,5];
+    expect(new Transform.negative(testArray).toEqual(3,2,1,0));
+  })
   // further tests on if class/objects work
 });
