+++ index.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const fs = require('fs');
+const Transform = require('./lib/transform');
 
 const regex = /(.bmp)$/g;
 
@@ -37,60 +38,6 @@ const writeNewFile = (inputFile, transformation) => {
     });
 };
 
-// test transformations
-
-class Transform{
-    constructor(colorArray){
-        this.colorArray = colorArray;
-    }
-
-    negative(){
-        let reverseCounter = this.colorArray.length - 1;
-        for(let i = 0; i < this.colorArray.length; i++){
-            this.colorArray[i] = reverseCounter;
-            reverseCounter --;
-        }
-    };
-
-    pinkHighlights(){
-        for(let i = 0; i < this.colorArray.length; i++){
-            if(this.colorArray[i] === 0){
-                this.colorArray[i] = 255;
-            }
-         }
-    };
-
-    allBlack(){
-        for(let i = 0; i < this.colorArray.length; i++){
-            this.colorArray[i] = 0;
-        }
-    };
-
-    allWhite(){
-        for(let i = 0; i < this.colorArray.length; i++){
-            this.colorArray[i] = 255;
-         }
-    };
-
-    neon1(){
-
-        for(let i = 0; i < this.colorArray.length; i+=3){
-            this.colorArray[i] = 0;
-            this.colorArray[i + 1] = 191;
-            this.colorArray[i + 2] = 255;
-        }
-    };
-
-    neon2(){
-
-        for(let i = 0; i < this.colorArray.length; i+=3){
-            this.colorArray[i] = 255;
-            this.colorArray[i + 1] = 0;
-            this.colorArray[i + 2] = 191;
-            this.colorArray[i + 3] = 0;
-        }
-    }
-}
 
 
 
@@ -141,10 +88,11 @@ const readFile = (file, transformation) => {
                 transformConstructor.neon2();
             }
             writeNewFile(data, transformation);
+            return colorArray;
     });
 
     if(!regex.test(`${file}`)) {
-        return'Not a .bmp file';
+        return 'Not a .bmp file';
     }
 };
 
@@ -156,4 +104,4 @@ readFile(file, transformation);
 
 console.log(file);
 
-module.exports = readFile, Transform;
+module.exports = readFile;
